tweak

+++ src/main/java/golemknights/tinkersgolem/mixinhelper/FireballModule_ShoulderWeaponModifierHook.java
@@ -6,29 +6,19 @@ import javax.annotation.Nullable;
 
 import dev.xkmc.modulargolems.content.entity.metalgolem.MetalGolemEntity;
 import dev.xkmc.modulargolems.content.item.ranged.CannonPoseUtil;
-import dev.xkmc.modulargolems.content.item.ranged.IShoulderWeapon;
-import dev.xkmc.modulargolems.util.TNTLauncher;
-import golemknights.tinkersgolem.hooks.ShoulderWeaponModifierHook;
-import net.minecraft.sounds.SoundEvents;
-import net.minecraft.util.RandomSource;
+import golemknights.tinkersgolem.library.hooks.ShoulderWeaponModifierHook;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.item.PrimedTnt;
-import net.minecraft.world.entity.projectile.LargeFireball;
-import net.minecraft.world.entity.projectile.SmallFireball;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.ProjectileWeaponItem;
-import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.phys.Vec3;
-import net.minecraftforge.common.ForgeHooks;
 import slimeknights.tconstruct.library.modifiers.ModifierEntry;
 import slimeknights.tconstruct.library.modifiers.ModifierHooks;
 import slimeknights.tconstruct.library.modifiers.fluid.entity.DamageFluidEffect.DamageTypePair;
 import slimeknights.tconstruct.library.modifiers.hook.build.ConditionalStatModifierHook;
 import slimeknights.tconstruct.library.tools.capability.EntityModifierCapability;
 import slimeknights.tconstruct.library.tools.capability.PersistentDataCapability;
-import slimeknights.tconstruct.library.tools.helper.ModifierUtil;
 import slimeknights.tconstruct.library.tools.helper.ToolDamageUtil;
 import slimeknights.tconstruct.library.tools.nbt.IToolStackView;
 import slimeknights.tconstruct.library.tools.nbt.ModDataNBT;
@@ -78,8 +68,8 @@ public interface FireballModule_ShoulderWeaponModifierHook extends ShoulderWeapo
                         return;
                     Vec3 pos = CannonPoseUtil.FLAME_THROWER.getOrigin(entity, hand);
 
-                    Vec3 dst = target.position().add((double) 0.0F, (double) (target.getBbHeight() / 2.0F),
-                            (double) 0.0F);
+                    Vec3 dst = target.position().add(0.0F, target.getBbHeight() / 2.0F,
+                            0.0F);
                     Vec3 dir = dst.subtract(pos).normalize();
 
                     Level level = entity.level();
